Moved utils into new module
