+++ app/src/main/java/com/upc/fib/racopocket/NotificationsMainMenu.java
@@ -15,7 +15,9 @@ import android.widget.ImageButton;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 
+import com.upc.fib.racopocket.Utils.Constants;
 import com.upc.fib.racopocket.Utils.FileUtils;
+import com.upc.fib.racopocket.Utils.TokensStorageUtils;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -51,8 +53,8 @@ public class NotificationsMainMenu extends Fragment
         progressBar = (ProgressBar) rootView.findViewById(R.id.progressBarNotifications);
         expListViewNotifications = (ExpandableListView) rootView.findViewById(R.id.expListViewNotifications);
 
-        String token = TokensStorageHelpers.recoverTokens(getContext().getApplicationContext(), "OAUTH_TOKEN");
-        String secret = TokensStorageHelpers.recoverTokens(getContext().getApplicationContext(), "OAUTH_TOKEN_SECRET");
+        String token = TokensStorageUtils.recoverTokens(getContext().getApplicationContext(), "OAUTH_TOKEN");
+        String secret = TokensStorageUtils.recoverTokens(getContext().getApplicationContext(), "OAUTH_TOKEN_SECRET");
         consumer.setTokenWithSecret(token, secret);
 
         return rootView;
