+++ app/src/main/java/com/upc/fib/racopocket/Utils/Constants.java
@@ -1,4 +1,4 @@
-package com.upc.fib.racopocket;
+package com.upc.fib.racopocket.Utils;
 
 public class Constants {
